zephyr: device: uart: Move UartIrq to own crate
Move this specific type into its own crate.
